Some fixes

+++ hikka/inline/form.py
@@ -179,7 +179,7 @@ class Form(InlineUnit):
 
         self._forms[form_uid]["chat"] = utils.get_chat_id(m)
         self._forms[form_uid]["message_id"] = m.id
-        if isinstance(message, Message):
+        if isinstance(message, Message) and message.out:
             await message.delete()
 
         if not any(
